Moved collections helpers to the collections package
