+++ composer.json
@@ -98,6 +98,7 @@
     },
     "autoload": {
         "files": [
+            "src/Illuminate/Collections/helpers.php",
             "src/Illuminate/Events/functions.php",
             "src/Illuminate/Foundation/helpers.php",
             "src/Illuminate/Support/helpers.php"
