+++ src/Illuminate/Collections/Arr.php
@@ -167,7 +167,7 @@ class Arr
     {
         if (is_null($callback)) {
             if (empty($array)) {
-                return Util::value($default);
+                return value($default);
             }
 
             foreach ($array as $item) {
@@ -181,7 +181,7 @@ class Arr
             }
         }
 
-        return Util::value($default);
+        return value($default);
     }
 
     /**
@@ -195,7 +195,7 @@ class Arr
     public static function last($array, callable $callback = null, $default = null)
     {
         if (is_null($callback)) {
-            return empty($array) ? Util::value($default) : end($array);
+            return empty($array) ? value($default) : end($array);
         }
 
         return static::first(array_reverse($array, true), $callback, $default);
@@ -286,7 +286,7 @@ class Arr
     public static function get($array, $key, $default = null)
     {
         if (! static::accessible($array)) {
-            return Util::value($default);
+            return value($default);
         }
 
         if (is_null($key)) {
@@ -298,14 +298,14 @@ class Arr
         }
 
         if (strpos($key, '.') === false) {
-            return $array[$key] ?? Util::value($default);
+            return $array[$key] ?? value($default);
         }
 
         foreach (explode('.', $key) as $segment) {
             if (static::accessible($array) && static::exists($array, $segment)) {
                 $array = $array[$segment];
             } else {
-                return Util::value($default);
+                return value($default);
             }
         }
